# docs/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/v{api_version}/ping":{"get":{"summary":"Ping","description":"Ping/pong for a quick health check","operationId":"ping_v_api_version__ping_get","parameters":[{"required":true,"schema":{"title":"Api Version"},"name":"api_version","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PingResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v{api_version}/repo/{owner}/{repo}/popularity":{"get":{"summary":"Endpoint Popularity","description":"Checking a repo's popularity.\n\n\"GitHub-Access-Token\" header is required\n\"X-Use-Caching: on\" header is optional","operationId":"endpoint_popularity_v_api_version__repo__owner___repo__popularity_get","parameters":[{"required":true,"schema":{"title":"Api Version","type":"integer"},"name":"api_version","in":"path"},{"required":true,"schema":{"title":"Owner","type":"string"},"name":"owner","in":"path"},{"required":true,"schema":{"title":"Repo","type":"string"},"name":"repo","in":"path"},{"required":false,"schema":{"title":"X-Use-Caching","type":"string"},"name":"x-use-caching","in":"header"},{"required":false,"schema":{"title":"Github-Access-Token","type":"string"},"name":"github-access-token","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RepoPopularityResponse"}}}},"400":{"description":"Access token is required","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Invalid access token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Repository not found or is private","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ErrorResponse":{"title":"ErrorResponse","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"PingResponse":{"title":"PingResponse","type":"object","properties":{"pong":{"title":"Pong","type":"boolean","default":true}}},"RepoPopularityResponse":{"title":"RepoPopularityResponse","required":["is_popular"],"type":"object","properties":{"is_popular":{"title":"Is Popular","type":"boolean"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}